--- src/components/MainContent/Chart/ChartOption/ChartOptionSubCategory/ChartOptionSubCategory.module.css
@@ -1,50 +1,0 @@
-.checkbox {
-  margin-top: 3rem;
-}
-.checkbox__container {
-  /* display: flex;
-  flex-direction: column; */
-  display: grid; /* fix flex gap */
-  gap: 1rem;
-}
-
-.input__container {
-  flex: 1 1 40%;
-}
-
-.input__container:hover .check {
-  border: 3px solid var(--checkbox-active);
-}
-.input__container:hover .label {
-  color: var(--checkbox-active);
-}
-
-.subtitle--time {
-  color: var(--blue);
-  margin-bottom: 0.2rem;
-}
-
-.check {
-  width: 2rem;
-  height: 2rem;
-  border-radius: 5px;
-  border: 3px solid var(--checkbox-passive);
-}
-
-.input:checked ~ .check {
-  background-color: var(--checkbox-active);
-  border: 3px solid var(--checkbox-active);
-}
-
-/* 
-"+" only select adjacent sibling
-use two "+" to select non-adjacent sibling 
-*/
-.input:checked + .check + .label {
-  color: var(--checkbox-active);
-}
-
-.label {
-  font-size: 1.6rem;
-  color: var(--checkbox-passive);
-}
